requirements.txt
requests
openai
pandas
numpy
scikit-learn
python-dateutil
Pillow
